+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- display registry and rpc over multichain settings in `/info` endpoint
+- display uniswap contract addresses in `/info` endpoint
 
 ## [7.5.0] - 2022-11-10
 ### Changed